ai_quiz_backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
openai>=1.6.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0